refactor: change column types of measuring data from DOUBLE to INT

+++ migrations/Version20260124110742.php
@@ -20,10 +20,10 @@ final class Version20260124110742 extends AbstractMigration
             CREATE TABLE sensor_data (
               id INT AUTO_INCREMENT NOT NULL,
               node_uuid VARCHAR(36) NOT NULL,
-              temperature DOUBLE PRECISION NOT NULL,
-              humidity DOUBLE PRECISION NOT NULL,
-              pressure DOUBLE PRECISION NOT NULL,
-              carbon_dioxide DOUBLE PRECISION DEFAULT NULL,
+              temperature INT NOT NULL,
+              humidity INT NOT NULL,
+              pressure INT NOT NULL,
+              carbon_dioxide INT DEFAULT NULL,
               measured_at DATETIME NOT NULL,
               created_at DATETIME NOT NULL,
               PRIMARY KEY(id)
